% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{CandySet}
\alias{CandySet}
\title{Candy Set}
\format{An object of class \code{data.frame} with 1842 rows and 4 columns.}
\usage{
CandySet
}
\description{
A dataset intended for use in the consideration set model, with the following
parameters: \itemize{ \item{Number of rows - 6 unique candies for 307
participants to choose one from, total of 1842 observations.} \item{Chosen -
Indicator of whether or not a participant chose a candy, 0 = "Not chosen" and
1 = "Chosen".} \item{Rank - Numerical value representing how each participant
ranks each of 6 candies relative to one another, with higher values meaning
higher favorability. Used to inform consideration step.} \item{SEQ -
Anonymized covariate used to inform consideration step of model.} \item{Rate
- Numerical value representing how each participant rates each of 6 candies
in more objective manner, with higher values meaning higher favorability.
Used to inform selection step.} }
}
\keyword{datasets}
