% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{CFS6_2}
\alias{CFS6_2}
\title{Consideration Set Simulation 4}
\format{An object of class \code{data.frame} with 30000 rows and 5 columns.}
\usage{
CFS6_2
}
\description{
A dataset intended for use in the consideration set model, with the following
parameters: \itemize{ \item{Number of rows - 6 unique options for 5000 
persons to choose from, total of 30000 observations.} \item{V1 - Indicator of
how many times an option was chosen, extending \code{\link{CFS6_1}} by 
including multiplicities.} \item{V2,V3,V4 - 3 covariates used to inform 
consideration and selection steps of model.} \item{V5 - Additional parameter 
indicating whether or not an item is available for consideration, 0 = 
"Unavailable" and 1 = "Available". In this simulation, 10\% of all items are 
unavailable.} }
}
\details{
Here is output we can expect from running the model on this data: \itemize{
\item{BetaConsid = [2, 0, -1]} \item{BetaSelect = [0, -.5, 1]}
\item{ASCsConsid = [.2, .4, .6, .8, 1, 1.2]} \item{ASCsSelect = [-1.25, -1,
-.75, -.5, -.25, 0]} }
}
\keyword{datasets}
