% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangePoint.R
\name{change.normal}
\alias{change.normal}
\title{Changepoint Model with Normal Distribution}
\usage{
change.normal(v, pre = NA, type = "MeanVar", know.mean = FALSE, mu = NA)
}
\arguments{
\item{v}{A vector that contains response variable of interest.}

\item{pre}{Numerical value that represents preselected changepoint for any
model, should the user wish to test a specific point in data; NA by
default.}

\item{type}{Choice between "Mean", "Var", and "MeanVar", regarding the
parameter(s) for which you wish to test significant change in value;
"MeanVar" by default.}

\item{know.mean}{(Valid only when \code{type="Var"}) Boolean variable, i.e.
TRUE or FALSE, indicating whether or not you wish to set a value for "true
mean" of data; FALSE by default.}

\item{mu}{(Valid only when \code{type="Var"} and \code{know.mean=TRUE})
Numerical value that represents the "true mean" of data; NA by default.}
}
\value{
List of the following: \itemize{ \item{Table containing location(s)
  of changepoint(s) -- either preselected or determined rigorously -- ranked
  by significance and associated significance value(s)} \item{Numerical value
  representing the MBIC threshold by which to base significance of a
  changepoint, i.e. penalty values LARGER than MBIC threshold indicate
  significance.} }
}
\description{
The changepoint model is used to test if there is a significant change in
mean and/or variance of a variable (assumed to have normal distribution),
over the course of some threshold, typically time or sequential order.
}
\examples{
a = rnorm(100,0,25/100)
b = rep(0,100)
for (i in 1:100) {
  b[i] = rnorm(1,0,(i/2+25)/100)
}
c = rnorm(100,0,75/100)
d = rep(0,100)
for (i in 1:100) {
  d[i] = rnorm(1,0,(75-i/2)/100)
}
e = rnorm(100,0,25/100)
v = c(a,b,c,d,e)

change.normal(v, type="Var", know.mean=TRUE, mu=0)
change.normal(v, type="Var", know.mean=TRUE, mu=0, pre=200)

}
