% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{dating}
\alias{dating}
\title{Dating}
\format{An object of class \code{list} of length 4: \itemize{ \item{stage1 -
  An object of class \code{data.frame} with 50000 rows and 8 columns.}
  \item{stage2 - An object of class \code{data.frame} with 26187 rows and 8
  columns.} \item{stage3 - An object of class \code{data.frame} with 9813
  rows and 8 columns.} \item{true_param - An object of class
  \code{data.frame} with 5 rows and 16 columns.} }}
\usage{
dating
}
\description{
A dataset intended for use in the discrete heterogeneity model, with the
following parameters: \itemize{ \item{respid - Unique ID for 696 persons in
the data.} \item{browsed - Choice variable representing first stage,
indicating whether or not a person browsed a profile; 0 = "Not browsed", 1 =
"Browsed".} \item{wrote - Choice variable representing second stage,
indicating whether or not a person wrote to a profile; 0 = "Not written to",
1 = "Written to".} \item{agedif - Covariate representing difference in age
between person and profile.} }
}
\keyword{datasets}
