% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCmods-package.r
\docType{package}
\name{DCmods}
\alias{DCmods}
\alias{DCmods-package}
\title{DCmods.}
\description{
This package contains various statistical models, collected primarily for
usage in marketing and social research on choice behavior data. The models
included are the changepoint model, the consideration set model, and the
discrete heterogeneity model.
}
\details{
For details on each model, please refer to the help pages of each function,
as well as a supplemental manual, "Better Understanding DCmods," the URL of
which will be included at a later date.
}
\author{
Elizabeth Bruch, Fred Feinberg, Cameron Hollingshead, Yue Wang,
  Yujing Zhou
}
