% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentStage.R
\name{LatentStage}
\alias{LatentStage}
\title{Discrete heterogeneity model with multiple stages}
\usage{
LatentStage(nclass, ...)
}
\arguments{
\item{nclass}{Number of classes for all subjects, determining how many groups
of different parameters will be obtained}

\item{...}{\itemize{\item{X1, X2, X3... - Matrix of covariates for 1st, 2nd,
3rd stage, respectively.} \item{y1, y2, y3... - Vector of responses for 1st,
2nd, 3rd stage, respectively.} \item{id1, id2, id3... - Vector of subject
IDs for 1st, 2nd, 3rd stage, respectively. The dimensions of these arguments
must be the same, i.e. row(X1) = length(y1) = length(id1), as they determine
the number of stages in the model.}}}
}
\value{
A list object containing both arguments and results: \itemize{
 \item{\code{lambda} - The estimate of class proportions, which sum up to 1.}
 \item{\code{beta} - Estimates, SEs and p-values for all linear parameters.}
 \item{\code{posteriorz} - List of probabilities of each subject belonging to
 a specific group, with class ID estimated by determining which one maximizes
 said probability.} \item{\code{all.loglik} - List of log-likelihood
 estimates in each iteration.} \item{\code{y} - List of responses for all
 stages.} \item{\code{id} - List of subject IDs for all stages.}
 \item{\code{x} - List of covariates for all stages.} \item{\code{AIC} -
 Numerical AIC value for the current model.} \item{\code{BIC} - Numerical BIC
 value for the current model.} \item{\code{runtime} - Elapsed time in fitting
 the model.}}
}
\description{
The discrete heterogeneity model is primarily applicable to the scenario in
which each subject is assigned a latent class, and fixed parameters are
applied across those subjects in the same class. This particular function
allows for more than one response, thus appropriate for constructing a model
with multiple stages. Although each response is typically nested within one
another, this property is not required, in order to fit the model.
}
\examples{
\donttest{
data(threestage)
attach(threestage)
mod <- LatentStage(5, X1=stage1[, 4:7],
       y1=stage1$Y1, id1=stage1$Person,
       X2=stage2[, 4:7],
       y2=stage2$Y2, id2=stage2$Person,
       X3=stage3[, 4:7],
       y3=stage3$Y3, id3=stage3$Person)

data(dating)
attach(dating)
nonmiss <- !is.na(wrote)
mod <- LatentStage(3, y1 = browsed, y2 = wrote[nonmiss],
                   id1 = respid, id2 = respid[nonmiss],
                   X1 = agedif, X2 = agedif[nonmiss])
}
}
