% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se_outer}
\alias{se_outer}
\title{Standard errors of parameters and proportions}
\usage{
se_outer(out)
}
\arguments{
\item{out}{A list object obtained directly from output of 
\code{\link{LatentStage}}}
}
\value{
A list object with two attributes: \itemize{\item{\code{se.beta} - 
  Outer-product SE of linear parameters.} \item{\code{se.lambda} - 
  Outer-product SE of class proportions. Note that the 
  \code{\link{LatentStage}} function does not provide this parameter.}}
}
\description{
Calculates outer-product of standard errors for a discrete heterogeneity 
model's estimate of linear parameters and class proportions. It is worth 
noting that, when the covariates of a given model have high linear 
correlation, the function will be unable to calculate this outer-product.
}
