% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{SnackSet}
\alias{SnackSet}
\title{Snack Set}
\format{An object of class \code{data.frame} with 2376 rows and 14 columns.}
\usage{
SnackSet
}
\description{
A dataset intended for use in the consideration set model, with the following
parameters: \itemize{ \item{Number of rows - 12 unique snacks for 198
participants to choose two from, total of 2376 observations.} \item{Chosen -
Indicator of whether or not a participant chose a snack, 0 = "Not chosen" and
1 = "Chosen".} \item{RankCent - Numerical value representing how each
participant ranks each of 12 snacks relative to one another, with higher
values meaning higher favorability. Used to inform consideration step.}
\item{D1,...,D11 - 11 anonymized covariates used to inform consideration and
selection steps of model.} \item{PrRateStd - Numerical value representing how
each participant rates each of 12 snacks in more objective manner, with
higher values meaning higher favorability. Used to inform selection step.} }
}
\keyword{datasets}
