% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDataFiles.R
\docType{data}
\name{threestage}
\alias{threestage}
\title{Three-Stage Simulation}
\format{An object of class \code{list} of length 4.}
\usage{
threestage
}
\description{
A dataset intended for use in the discrete heterogeneity model, with the 
following parameters: \itemize{ \item{Person - Unique ID for 250 persons in 
the data.} \item{Class - Arbitrary class (out of 5) for all persons.} 
\item{Obs - Unique ID for 200 sequenced observations of each person in the 
data.} \item{XA1,...,XB1,...,XC1,... - 3 unique covariates for each person at
each stage.} \item{Y1,...,Y3 - Response value for each person at each stage.}
\item{true_param - Table of true values of class proportions and 
stage-dependent covariates.} }
}
\keyword{datasets}
